Completed graphing of values
Graphing of records now works with user-defined values for graph
titles, axes labels (using a dict) and filenames.

For example, if passed the following tuple:
title_pattern = ( [0, 'test in', 2, '\n', 4, ' (', 6, ')'],
        ["Type", "BoreholeID", "Farm", "Owner"] )
will then be combined to use the Type, BoreholeID, Farm and Owner
of each record, filling those values in, in order within the
overall pattern.

At present this is used for generating graph titles and graph
filenames using values from the record.

I anticipate using this for marker text and report pages as well.