Added code for generating borehole reports.
Reports now work. Generates a new html page for each report,
according to a template given in the same way as the other text
generated.

If a folder with photos is found, a link is provided. If a graph
for a given test is found, it is linked in as an image.

This should be linked to from the pop-up itself.