Full functionality of old version replicated
The old version that this is a rewrite of appears to be obsolete
now, since this version appears to be functioning to the same
standard: markers are generated with a link to a report.

The report will have a link to pictures (if present), graphs (if
present). If only one value is present, then it is assumed that
the test was only performed at surface, which is explicitly stated
in the report.

I have also added a "verbose mode". This will essentially print out
pretty much everything that each function is doing. There will need
to be a better way of doing this, but piping output to a file is a
workable solution for now. This defaults to off, so needs to be
explicitly requested in the function call. Most functions have this,
largely for my own debugging purposes.
An example of the program running on the full test dataset is
included as "sample_run".